moved components related to the assessment page into their own folder to make things a little clearer. Vscode doesn't seem to like it atm though...
